% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_definitions.R
\name{cluster_definitions}
\alias{cluster_definitions}
\title{Cluster Definitions}
\usage{
cluster_definitions()
}
\value{
Dataframe of cluster definitions
}
\description{
Get explanations of cluster teams, statistical similarities
}
