% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_team_clusters.R
\name{current_team_clusters}
\alias{current_team_clusters}
\title{Current Team Cluster Groups}
\usage{
current_team_clusters()
}
\value{
Dataframe of team statistics and cluster group
}
\description{
Get all teams current advanced statistics and clusters
}
