% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentTeamStatus.R
\name{currentTeamStatus}
\alias{currentTeamStatus}
\title{Get current statistics for every team}
\usage{
currentTeamStatus()
}
\value{
Dataframe of team statistics
}
\description{
Gets the last ten games advanced statistics for
every team up to date
}
