% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkPoints.R
\name{dkPoints}
\alias{dkPoints}
\title{Calculate DK fantasy points}
\usage{
dkPoints(pts, rebs, ast, stl, blk, three, tov)
}
\arguments{
\item{pts}{Points scored}

\item{rebs}{Rebounds}

\item{ast}{Assists}

\item{stl}{Steals}

\item{blk}{Blocks}

\item{three}{3 pointers made}

\item{tov}{turnovers}
}
\value{
dk fantasy points
}
\description{
Use a players box score statistics to get their 
dk fantasy points
}
