% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_ten_games.R
\name{last_ten_games}
\alias{last_ten_games}
\title{Calculate last 10 games statistics}
\usage{
last_ten_games(games, team, game)
}
\arguments{
\item{games}{Dataframe of games taken from the nbastatR package}

\item{team}{String of team abbreviation to get their statistics}

\item{game}{Integer for team's game number of season}
}
\value{
Dataframe with team advanced statistics
}
\description{
Calculate a teams advanced statistics in their 
last ten games
}
