% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_past_games.R
\name{player_past_games}
\alias{player_past_games}
\title{Get player statistics for all games}
\usage{
player_past_games(player)
}
\arguments{
\item{player}{String of player name}
}
\value{
Dataframe of all player games
}
\description{
Retrieve player fantasy points and opponent statistics
for all games in 2019
}
