% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_games.R
\name{similar_games}
\alias{similar_games}
\title{Player games against similar opponents}
\usage{
similar_games(player, team)
}
\arguments{
\item{player}{String of player name}

\item{team}{String of upcoming team}
}
\value{
Dataframe of games against similar teams
}
\description{
Get 5 most recent games the player had against similar
teams as their upcoming games
}
